% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aapltrend.R
\docType{data}
\name{aaplt}
\alias{aaplt}
\title{Run command}
\format{An object of class \code{plotly} (inherits from \code{htmlwidget}) of length 9.}
\usage{
aaplt
}
\arguments{
\item{aaplt}{we can see fast the Apple stocks trend in last 90 days (title, axis, animation; you can change parameters to have a different visual, slower/faster animation, different values for axes, ecc...)}
}
\value{
closing of Apple stocks in the last 90 days; press "Play" to have funny animation of increase or decrease treand
}
\description{
Run command
}
\keyword{datasets}
