% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aapltrend.R
\name{accumulate_by}
\alias{accumulate_by}
\title{Levels stocks}
\usage{
accumulate_by(dat, var)
}
\arguments{
\item{accumulate_by}{is the lavels of the stocks by steps}
}
\value{
function(dat, var) {
var <- lazyeval::f_eval(var, dat)
lvls <- plotly:::getLevels(var)
dats <- lapply(seq_along(lvls), function(x) {
 cbind(dat[var %in% lvls[seq(1, x)], ], frame = lvls[[x]])
})
dplyr::bind_rows(dats)
}
}
\description{
Levels stocks
}
