% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{create_graph}
\alias{create_graph}
\title{Graph building}
\usage{
create_graph(accumulate_by, df, aaplt)
}
\arguments{
\item{accumulate_by}{accumulate by var, levels, dats}

\item{df}{data frame accumalte by ID}

\item{aaplt}{build of the graphic}
}
\value{
Animation graphic, with "play" bottom; write aaplt to Viewer
}
\description{
Graph building
}
