% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_acquisition}
\alias{data_acquisition}
\title{Data acquisition}
\usage{
data_acquisition(df)
}
\arguments{
\item{df}{data frame (apple stocks)}
}
\value{
total storic trend of Apple stock (source of data Yahoo: Date AAPL.Open AAPL.High AAPL.Low AAPL.Close AAPL.Volume AAPL.Adjusted)
}
\description{
Data acquisition
}
