% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aapltrend.R
\docType{data}
\name{df}
\alias{df}
\title{Call the storic trend}
\format{An object of class \code{data.frame} with 465 rows and 9 columns.}
\usage{
df
}
\arguments{
\item{1.df}{is a comand to call the storic trend of stocks with cod AAPL}
}
\value{
all trend storic present in yahoo database order by rows and contains: Date, AAPL.Open, AAPL.High, AAPL Low, AAPL Close, AAPL.Volume, AAPL.Adjusted
}
\description{
Call the storic trend
}
\keyword{datasets}
