% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_days.R
\name{last_days}
\alias{last_days}
\title{Last 90 days of trend}
\usage{
last_days(df)
}
\arguments{
\item{df}{data frame with last 90 rows}
}
\value{
last 90 days of storic trend of Apple stock (source of data Yahoo Date AAPL.Open AAPL.High AAPL.Low AAPL.Close AAPL.Volume AAPL.Adjusted)
}
\description{
Last 90 days of trend
}
