% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accumlate_byID.R
\name{order_trend}
\alias{order_trend}
\title{Order and setup}
\usage{
order_trend(order)
}
\arguments{
\item{order}{}
}
\value{
last 90 days of storic trend of Apple stock (source of data Yahoo Date AAPL.Open AAPL.High AAPL.Low AAPL.Close AAPL.Volume AAPL.Adjusted whit ID order form 1 to 90); accumulate by dat and var.
}
\description{
Order and setup
}
