% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aapltrend.R
\docType{data}
\name{x}
\alias{x}
\title{Data Acquisition}
\format{An object of class \code{character} of length 1.}
\usage{
x
}
\arguments{
\item{x}{"AAPL" stocks code, "yahoo" data source}
}
\description{
Data Acquisition
}
\keyword{datasets}
