% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_acquisition}
\alias{data_acquisition}
\title{Data acquisition}
\usage{
data_acquisition(data_acquisition)
}
\arguments{
\item{data_acquisition}{data acquisition, cod stocks: AAPL, source: yahoo}
}
\value{
total storic trend of Apple stock (source of data Yahoo Date AAPL.Open AAPL.High AAPL.Low AAPL.Close AAPL.Volume AAPL.Adjusted)
}
\description{
Data acquisition
}
\examples{
execute print(data_acquisition()) to see the list
}
